% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_neiss.R
\name{cluster_neiss}
\alias{cluster_neiss}
\title{Cluster \code{neiss1617}}
\usage{
cluster_neiss(narrative_regex = ".*", dist_method = "euclidean",
  hclust_method = "complete")
}
\arguments{
\item{narrative_regex}{A regular expression for filtering \code{neiss1617$Narrative_1}.}

\item{dist_method}{The method to use for calculating distance.}

\item{hclust_method}{The method to use for hierarhcical clustering.}
}
\value{
A \code{hclust} object.
}
\description{
The user can cluster the age groups in \code{neiss1617} after filtering the
\code{Narrative_1} using a regular expression.
}
