% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neiss1617.R
\docType{data}
\name{neiss1617}
\alias{neiss1617}
\title{The National Electronic Injury Surveillance System data from 2016-17}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 762104 rows and 24 columns.}
\usage{
neiss1617
}
\description{
A data frame with all of the NEISS data combined. This includes the human-
readable descriptions of products, diagnoses, and bodyparts. Additionally,
the ages of infants has been adjusted into years instead of 200 added to the
number of months.
}
\keyword{datasets}
